supplemental-analysis/combineTCR-benchmark/py-benchmark/sc-dandelion-benchmark/requirements.txt
adjustText==1.3.0
airr==1.5.1
anndata==0.11.3
array_api_compat==1.11.2
biopython==1.85
changeo==1.3.0
contourpy==1.3.2
cycler==0.12.1
Distance==0.1.3
fonttools==4.57.0
h5py==3.13.0
igraph==0.11.8
joblib==1.4.2
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
leidenalg==0.10.2
llvmlite==0.44.0
matplotlib==3.10.1
mizani==0.13.1
more-itertools==10.6.0
natsort==8.4.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nxviz==0.7.6
packaging==24.2
palettable==3.3.3
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotnine==0.14.5
polyleven==0.9.0
presto==0.7.4
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
sc-dandelion==0.5.4
scanpy==1.11.0
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
session-info2==0.1.2
setuptools==78.1.0
setuptools-scm==8.2.0
six==1.17.0
statsmodels==0.14.4
texttable==1.7.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
yamlordereddictloader==0.4.2
